+++ src/SFA.DAS.QnA.Api/Startup.cs
@@ -20,6 +20,7 @@ using Microsoft.Extensions.Options;
 using SFA.DAS.Qna.Data;
 using SFA.DAS.QnA.Api.Infrastructure;
 using SFA.DAS.QnA.Application;
+using SFA.DAS.QnA.Application.Commands.SetPageAnswers;
 using SFA.DAS.QnA.Application.Validators;
 using SFA.DAS.QnA.Configuration.Config;
 using SFA.DAS.QnA.Configuration.Infrastructure;
@@ -84,6 +85,7 @@ namespace SFA.DAS.QnA.Api
 
             services.RegisterAllTypes<IValidator>(new[] { typeof(IValidator).Assembly });
             services.AddTransient<IValidatorFactory, ValidatorFactory>();
+            services.AddTransient<IAnswerValidator, AnswerValidator>();
             
             services.AddAutoMapper(typeof(HandlerResponse<>).Assembly);
             services.AddMediatR(AppDomain.CurrentDomain.GetAssemblies());
